+++ scripts/generate_project.py
@@ -3,6 +3,8 @@
 import hashlib
 def uid(s): return hashlib.md5(s.encode()).hexdigest()[:24].upper()
 swift=["SuperkeysApp","Updater","WhatsNew","Onboarding","AppState","HyperEventTap","CapsLock","KeyCodes","WindowManager","WindowArranger","AXWindow","SpaceManager","SkyLightBridge","AppLauncher","BindingsStore","Permissions","SettingsRootView","ShortcutsView","Components","SettingsWindowController","CheatSheet","ConfigFile","MenuBarContent"]
+# Unit tests, hosted in the app (which skips its startup under XCTest).
+tests=["ConfigFileTests","KeystrokeTests","KeyHandlingTests","KeyGroupsTests"]
 fw=["AppKit","SwiftUI","ApplicationServices","ServiceManagement","UniformTypeIdentifiers","Combine","Carbon","CoreGraphics"]
 L=[]
 a=L.append
@@ -12,7 +14,18 @@ for n in swift: a(f"\t\t{uid('bf'+n)} /* {n}.swift in Sources */ = {{isa = PBXBu
 a(f"\t\t{uid('bfassets')} /* Assets.xcassets in Resources */ = {{isa = PBXBuildFile; fileRef = {uid('frassets')} /* Assets.xcassets */; }};")
 for f in fw: a(f"\t\t{uid('bffw'+f)} /* {f}.framework in Frameworks */ = {{isa = PBXBuildFile; fileRef = {uid('frfw'+f)} /* {f}.framework */; }};")
 a(f"\t\t{uid('bfsparkle')} /* Sparkle in Frameworks */ = {{isa = PBXBuildFile; productRef = {uid('sparkleprod')} /* Sparkle */; }};")
+for n in tests: a(f"\t\t{uid('tbf'+n)} /* {n}.swift in Sources */ = {{isa = PBXBuildFile; fileRef = {uid('tfr'+n)} /* {n}.swift */; }};")
 a("/* End PBXBuildFile section */\n")
+a(f"""/* Begin PBXContainerItemProxy section */
+\t\t{uid('tproxy')} /* PBXContainerItemProxy */ = {{
+\t\t\tisa = PBXContainerItemProxy;
+\t\t\tcontainerPortal = {uid('project')} /* Project object */;
+\t\t\tproxyType = 1;
+\t\t\tremoteGlobalIDString = {uid('target')};
+\t\t\tremoteInfo = Superkeys;
+\t\t}};
+/* End PBXContainerItemProxy section */
+""")
 a("/* Begin PBXFileReference section */")
 a(f"\t\t{uid('product')} /* Superkeys.app */ = {{isa = PBXFileReference; explicitFileType = wrapper.application; includeInIndex = 0; path = Superkeys.app; sourceTree = BUILT_PRODUCTS_DIR; }};")
 for n in swift: a(f"\t\t{uid('fr'+n)} /* {n}.swift */ = {{isa = PBXFileReference; lastKnownFileType = sourcecode.swift; path = {n}.swift; sourceTree = \"<group>\"; }};")
@@ -21,22 +34,28 @@ a(f"\t\t{uid('frplist')} /* Info.plist */ = {{isa = PBXFileReference; lastKnownF
 a(f"\t\t{uid('frxcconfig')} /* Signing.xcconfig */ = {{isa = PBXFileReference; lastKnownFileType = text.xcconfig; path = Signing.xcconfig; sourceTree = \"<group>\"; }};")
 a(f"\t\t{uid('frent')} /* Superkeys.entitlements */ = {{isa = PBXFileReference; lastKnownFileType = text.plist.entitlements; path = Superkeys.entitlements; sourceTree = \"<group>\"; }};")
 for f in fw: a(f"\t\t{uid('frfw'+f)} /* {f}.framework */ = {{isa = PBXFileReference; lastKnownFileType = wrapper.framework; name = {f}.framework; path = System/Library/Frameworks/{f}.framework; sourceTree = SDKROOT; }};")
+a(f"\t\t{uid('testproduct')} /* SuperkeysTests.xctest */ = {{isa = PBXFileReference; explicitFileType = wrapper.cfbundle; includeInIndex = 0; path = SuperkeysTests.xctest; sourceTree = BUILT_PRODUCTS_DIR; }};")
+for n in tests: a(f"\t\t{uid('tfr'+n)} /* {n}.swift */ = {{isa = PBXFileReference; lastKnownFileType = sourcecode.swift; path = {n}.swift; sourceTree = \"<group>\"; }};")
 a("/* End PBXFileReference section */\n")
 a("/* Begin PBXFrameworksBuildPhase section */")
 a(f"\t\t{uid('fwphase')} /* Frameworks */ = {{\n\t\t\tisa = PBXFrameworksBuildPhase;\n\t\t\tbuildActionMask = 2147483647;\n\t\t\tfiles = (")
 for f in fw: a(f"\t\t\t\t{uid('bffw'+f)} /* {f}.framework in Frameworks */,")
 a(f"\t\t\t\t{uid('bfsparkle')} /* Sparkle in Frameworks */,")
 a("\t\t\t);\n\t\t\trunOnlyForDeploymentPostprocessing = 0;\n\t\t};")
+a(f"\t\t{uid('tfwphase')} /* Frameworks */ = {{\n\t\t\tisa = PBXFrameworksBuildPhase;\n\t\t\tbuildActionMask = 2147483647;\n\t\t\tfiles = (\n\t\t\t);\n\t\t\trunOnlyForDeploymentPostprocessing = 0;\n\t\t}};")
 a("/* End PBXFrameworksBuildPhase section */\n")
 a("/* Begin PBXGroup section */")
-a(f"\t\t{uid('gmain')} = {{\n\t\t\tisa = PBXGroup;\n\t\t\tchildren = (\n\t\t\t\t{uid('frxcconfig')} /* Signing.xcconfig */,\n\t\t\t\t{uid('gsrc')} /* Superkeys */,\n\t\t\t\t{uid('gfw')} /* Frameworks */,\n\t\t\t\t{uid('gprod')} /* Products */,\n\t\t\t);\n\t\t\tsourceTree = \"<group>\";\n\t\t}};")
+a(f"\t\t{uid('gmain')} = {{\n\t\t\tisa = PBXGroup;\n\t\t\tchildren = (\n\t\t\t\t{uid('frxcconfig')} /* Signing.xcconfig */,\n\t\t\t\t{uid('gsrc')} /* Superkeys */,\n\t\t\t\t{uid('gtests')} /* SuperkeysTests */,\n\t\t\t\t{uid('gfw')} /* Frameworks */,\n\t\t\t\t{uid('gprod')} /* Products */,\n\t\t\t);\n\t\t\tsourceTree = \"<group>\";\n\t\t}};")
 a(f"\t\t{uid('gsrc')} /* Superkeys */ = {{\n\t\t\tisa = PBXGroup;\n\t\t\tchildren = (")
 for n in swift: a(f"\t\t\t\t{uid('fr'+n)} /* {n}.swift */,")
 a(f"\t\t\t\t{uid('frassets')} /* Assets.xcassets */,\n\t\t\t\t{uid('frplist')} /* Info.plist */,\n\t\t\t\t{uid('frent')} /* Superkeys.entitlements */,\n\t\t\t);\n\t\t\tpath = Superkeys;\n\t\t\tsourceTree = \"<group>\";\n\t\t}};")
+a(f"\t\t{uid('gtests')} /* SuperkeysTests */ = {{\n\t\t\tisa = PBXGroup;\n\t\t\tchildren = (")
+for n in tests: a(f"\t\t\t\t{uid('tfr'+n)} /* {n}.swift */,")
+a("\t\t\t);\n\t\t\tpath = SuperkeysTests;\n\t\t\tsourceTree = \"<group>\";\n\t\t};")
 a(f"\t\t{uid('gfw')} /* Frameworks */ = {{\n\t\t\tisa = PBXGroup;\n\t\t\tchildren = (")
 for f in fw: a(f"\t\t\t\t{uid('frfw'+f)} /* {f}.framework */,")
 a("\t\t\t);\n\t\t\tname = Frameworks;\n\t\t\tsourceTree = \"<group>\";\n\t\t};")
-a(f"\t\t{uid('gprod')} /* Products */ = {{\n\t\t\tisa = PBXGroup;\n\t\t\tchildren = (\n\t\t\t\t{uid('product')} /* Superkeys.app */,\n\t\t\t);\n\t\t\tname = Products;\n\t\t\tsourceTree = \"<group>\";\n\t\t}};")
+a(f"\t\t{uid('gprod')} /* Products */ = {{\n\t\t\tisa = PBXGroup;\n\t\t\tchildren = (\n\t\t\t\t{uid('product')} /* Superkeys.app */,\n\t\t\t\t{uid('testproduct')} /* SuperkeysTests.xctest */,\n\t\t\t);\n\t\t\tname = Products;\n\t\t\tsourceTree = \"<group>\";\n\t\t}};")
 a("/* End PBXGroup section */\n")
 a("/* Begin PBXNativeTarget section */")
 a(f"""\t\t{uid('target')} /* Superkeys */ = {{
@@ -58,6 +77,23 @@ a(f"""\t\t{uid('target')} /* Superkeys */ = {{
 \t\t\tproductName = Superkeys;
 \t\t\tproductReference = {uid('product')} /* Superkeys.app */;
 \t\t\tproductType = "com.apple.product-type.application";
+\t\t}};
+\t\t{uid('ttarget')} /* SuperkeysTests */ = {{
+\t\t\tisa = PBXNativeTarget;
+\t\t\tbuildConfigurationList = {uid('cltests')} /* Build configuration list for PBXNativeTarget "SuperkeysTests" */;
+\t\t\tbuildPhases = (
+\t\t\t\t{uid('tsrcphase')} /* Sources */,
+\t\t\t\t{uid('tfwphase')} /* Frameworks */,
+\t\t\t);
+\t\t\tbuildRules = (
+\t\t\t);
+\t\t\tdependencies = (
+\t\t\t\t{uid('tdep')} /* PBXTargetDependency */,
+\t\t\t);
+\t\t\tname = SuperkeysTests;
+\t\t\tproductName = SuperkeysTests;
+\t\t\tproductReference = {uid('testproduct')} /* SuperkeysTests.xctest */;
+\t\t\tproductType = "com.apple.product-type.bundle.unit-test";
 \t\t}};""")
 a("/* End PBXNativeTarget section */\n")
 a(f"""/* Begin PBXProject section */
@@ -85,6 +121,7 @@ a(f"""/* Begin PBXProject section */
 \t\t\tprojectRoot = "";
 \t\t\ttargets = (
 \t\t\t\t{uid('target')} /* Superkeys */,
+\t\t\t\t{uid('ttarget')} /* SuperkeysTests */,
 \t\t\t);
 \t\t}};
 /* End PBXProject section */
@@ -92,7 +129,11 @@ a(f"""/* Begin PBXProject section */
 a(f"/* Begin PBXResourcesBuildPhase section */\n\t\t{uid('respase')} /* Resources */ = {{\n\t\t\tisa = PBXResourcesBuildPhase;\n\t\t\tbuildActionMask = 2147483647;\n\t\t\tfiles = (\n\t\t\t\t{uid('bfassets')} /* Assets.xcassets in Resources */,\n\t\t\t);\n\t\t\trunOnlyForDeploymentPostprocessing = 0;\n\t\t}};\n/* End PBXResourcesBuildPhase section */\n")
 a(f"/* Begin PBXSourcesBuildPhase section */\n\t\t{uid('srcphase')} /* Sources */ = {{\n\t\t\tisa = PBXSourcesBuildPhase;\n\t\t\tbuildActionMask = 2147483647;\n\t\t\tfiles = (")
 for n in swift: a(f"\t\t\t\t{uid('bf'+n)} /* {n}.swift in Sources */,")
+a("\t\t\t);\n\t\t\trunOnlyForDeploymentPostprocessing = 0;\n\t\t};")
+a(f"\t\t{uid('tsrcphase')} /* Sources */ = {{\n\t\t\tisa = PBXSourcesBuildPhase;\n\t\t\tbuildActionMask = 2147483647;\n\t\t\tfiles = (")
+for n in tests: a(f"\t\t\t\t{uid('tbf'+n)} /* {n}.swift in Sources */,")
 a("\t\t\t);\n\t\t\trunOnlyForDeploymentPostprocessing = 0;\n\t\t};\n/* End PBXSourcesBuildPhase section */\n")
+a(f"/* Begin PBXTargetDependency section */\n\t\t{uid('tdep')} /* PBXTargetDependency */ = {{\n\t\t\tisa = PBXTargetDependency;\n\t\t\ttarget = {uid('target')} /* Superkeys */;\n\t\t\ttargetProxy = {uid('tproxy')} /* PBXContainerItemProxy */;\n\t\t}};\n/* End PBXTargetDependency section */\n")
 
 common="""\t\t\t\tALWAYS_SEARCH_USER_PATHS = NO;
 \t\t\t\tCLANG_ENABLE_MODULES = YES;
@@ -148,6 +189,20 @@ dev=t.replace("ASSETCATALOG_COMPILER_APPICON_NAME = AppIcon;","ASSETCATALOG_COMP
 assert dev != t
 for cfg,k,settings in (("Debug","tdebug",dev),("Release","trelease",t)):
     a(f"\t\t{uid(k)} /* {cfg} */ = {{\n\t\t\tisa = XCBuildConfiguration;\n\t\t\tbaseConfigurationReference = {uid('frxcconfig')} /* Signing.xcconfig */;\n\t\t\tbuildSettings = {{\n{settings}\n\t\t\t}};\n\t\t\tname = {cfg};\n\t\t}};")
+for cfg,k,host in (("Debug","ttdebug","Superkeys Dev.app/Contents/MacOS/Superkeys Dev"),("Release","ttrelease","Superkeys.app/Contents/MacOS/Superkeys")):
+    a(f"""\t\t{uid(k)} /* {cfg} */ = {{
+\t\t\tisa = XCBuildConfiguration;
+\t\t\tbuildSettings = {{
+\t\t\t\tBUNDLE_LOADER = "$(TEST_HOST)";
+\t\t\t\tCODE_SIGN_IDENTITY = "-";
+\t\t\t\tCODE_SIGN_STYLE = Manual;
+\t\t\t\tGENERATE_INFOPLIST_FILE = YES;
+\t\t\t\tPRODUCT_BUNDLE_IDENTIFIER = space.superkeys.tests;
+\t\t\t\tPRODUCT_NAME = SuperkeysTests;
+\t\t\t\tTEST_HOST = "$(BUILT_PRODUCTS_DIR)/{host}";
+\t\t\t}};
+\t\t\tname = {cfg};
+\t\t}};""")
 a("/* End XCBuildConfiguration section */\n")
 a(f"""/* Begin XCConfigurationList section */
 \t\t{uid('clproject')} /* Build configuration list for PBXProject "Superkeys" */ = {{
@@ -164,6 +219,15 @@ a(f"""/* Begin XCConfigurationList section */
 \t\t\tbuildConfigurations = (
 \t\t\t\t{uid('tdebug')} /* Debug */,
 \t\t\t\t{uid('trelease')} /* Release */,
+\t\t\t);
+\t\t\tdefaultConfigurationIsVisible = 0;
+\t\t\tdefaultConfigurationName = Release;
+\t\t}};
+\t\t{uid('cltests')} /* Build configuration list for PBXNativeTarget "SuperkeysTests" */ = {{
+\t\t\tisa = XCConfigurationList;
+\t\t\tbuildConfigurations = (
+\t\t\t\t{uid('ttdebug')} /* Debug */,
+\t\t\t\t{uid('ttrelease')} /* Release */,
 \t\t\t);
 \t\t\tdefaultConfigurationIsVisible = 0;
 \t\t\tdefaultConfigurationName = Release;
@@ -204,6 +268,12 @@ scheme=f'''<?xml version="1.0" encoding="UTF-8"?>
       </BuildActionEntries>
    </BuildAction>
    <TestAction buildConfiguration = "Debug" selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB" selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB" shouldUseLaunchSchemeArgsEnv = "YES">
+      <Testables>
+         <TestableReference skipped = "NO">
+            <BuildableReference BuildableIdentifier = "primary" BlueprintIdentifier = "{uid('ttarget')}" BuildableName = "SuperkeysTests.xctest" BlueprintName = "SuperkeysTests" ReferencedContainer = "container:Superkeys.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
+      </Testables>
    </TestAction>
    <LaunchAction buildConfiguration = "Debug" selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB" selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB" launchStyle = "0" useCustomWorkingDirectory = "NO" ignoresPersistentStateOnLaunch = "NO" debugDocumentVersioning = "YES" debugServiceExtension = "internal" allowLocationSimulation = "YES">
       <BuildableProductRunnable runnableDebuggingMode = "0">